+++ src/main/java/com/devagit/springbootstudy/request/comment/CommentRequest.java
@@ -12,12 +12,12 @@ import java.sql.Timestamp;
 public class CommentRequest {
     @Nullable //달아야함
     private Integer parentId;
-    private int postId;
+    private long postId;
     private int sorts;
     private String userId;
     private String content;
 
-    public CommentRequest(@Nullable int parentId, int postId, String userId, String content, int sorts) {
+    public CommentRequest(@Nullable int parentId, long postId, String userId, String content, int sorts) {
         this.parentId = parentId;
         this.sorts = sorts;
         this.postId = postId;
